move x86 implementations of simd_utils.h to util/arch/x86/
